Rc/add measure to map (#112)
vl-toggle-button

Added the new vl-toggle-button component, which can be used controlled and uncontrolled.

vl-mapactions

Moved the vl-mapactions project inside the vl-map component. Changes that are made to the functionalities of the files that were in the vl-mapactions project will from now on be available under vl-map in these changelogs.

vl-map 💥

This component now includes the vl-mapactions files, as these were moved into this project in this release (see previous list item).

Upgraded OpenLayers dependency to v6.14.1 and jsts to v2.8.1.

Added vl-map-action-controls as a wrapper for map action controls which will be shown horizontally on the top right of the map.

Added vl-map-measure-control to control the active state of the measure action. The control temporarily contains the text "Meten", but should change to a ruler icon when this becomes available.

Fixed a bug where the tooltip overlay wouldn't show on the map when using the measure option on draw actions.

Changed the existing measure action. The tooltip overlays use vl-pill to show the measurement now and are removable because of the closable functionality of the pill component. The measurement can now take place between two points only. The measure action can now optionally be controlled using the new vl-map-measure-control or controlling the active state of the measure action from outside the map.

Made a change in the functionality of the data-vl-default-active attribute for map actions. From now on, the default active action will always be the action with this attribute, instead of the first action in the list of actions on the map.

Made a change to the ESC functionality of the map. From now on, when an action is activate and the ESC key is pressed, its stop() implementation is triggered instead of deactivating the action and activating the default active action (with the old implementation as mentioned aboven which resulted in often re-activating the action that was already active). When there is no active action, the default active action is activated.

Changed the map event action-activated to vl-active-action-changed, which now gives back the information of which action was deactivated and which action was activated. This event is internally used to handle the active state of both actions and their controls.

Added the map event vl-layer-visible-changed, which provides the information of which layer had its visiblity changed and what the new state of that visibility is. This event is internally used to handle the active state of both actions and their controls.

vl-pill

Made the background of vl-pill transparent when it's used in vl-map.